Read Files

Reads multiple files from a directory.

read_files(.dir = ".", .pattern = "[.]rds$", .fun = readRDS,
  .recursive = FALSE, .hidden = FALSE, .silent = FALSE, ...)

Arguments

.dir
A string of the path to the directory.

.pattern
A string of the regular expression to use when searching for files. To include all files use ".*".

.fun
The function to use when reading each file. The first argument must be the name of the file.

.recursive
A flag indicating whether to recurse into subdirectories.

.hidden
A flag indicating whether to include hidden files.

.silent
A flag indicating whether error messages should be suppressed when trying to read a file.

...
Additional arguments passed to .fun

Value

A named list of the files.

See also

Examples

dir <- system.file("extdata", package = "filer")
read_files(dir)

#> $chickwts.rds
#>    weight      feed
#> 1     179 horsebean
#> 2     160 horsebean
#> 3     136 horsebean
#> 4     227 horsebean
#> 5     217 horsebean
#> 6     168 horsebean
#> 7     108 horsebean
#> 8     124 horsebean
#> 9     143 horsebean
#> 10    140 horsebean
#> 11    309   linseed
#> 12    229   linseed
#> 13    181   linseed
#> 14    141   linseed
#> 15    260   linseed
#> 16    203   linseed
#> 17    148   linseed
#> 18    169   linseed
#> 19    213   linseed
#> 20    257   linseed
#> 21    244   linseed
#> 22    271   linseed
#> 23    243   soybean
#> 24    230   soybean
#> 25    248   soybean
#> 26    327   soybean
#> 27    329   soybean
#> 28    250   soybean
#> 29    193   soybean
#> 30    271   soybean
#> 31    316   soybean
#> 32    267   soybean
#> 33    199   soybean
#> 34    171   soybean
#> 35    158   soybean
#> 36    248   soybean
#> 37    423 sunflower
#> 38    340 sunflower
#> 39    392 sunflower
#> 40    339 sunflower
#> 41    341 sunflower
#> 42    226 sunflower
#> 43    320 sunflower
#> 44    295 sunflower
#> 45    334 sunflower
#> 46    322 sunflower
#> 47    297 sunflower
#> 48    318 sunflower
#> 49    325  meatmeal
#> 50    257  meatmeal
#> 51    303  meatmeal
#> 52    315  meatmeal
#> 53    380  meatmeal
#> 54    153  meatmeal
#> 55    263  meatmeal
#> 56    242  meatmeal
#> 57    206  meatmeal
#> 58    344  meatmeal
#> 59    258  meatmeal
#> 60    368    casein
#> 61    390    casein
#> 62    379    casein
#> 63    260    casein
#> 64    404    casein
#> 65    318    casein
#> 66    352    casein
#> 67    359    casein
#> 68    216    casein
#> 69    222    casein
#> 70    283    casein
#> 71    332    casein
#> 
#> $mtcars.rds
#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> 
#> $trees.rds
#>    Girth Height Volume
#> 1    8.3     70   10.3
#> 2    8.6     65   10.3
#> 3    8.8     63   10.2
#> 4   10.5     72   16.4
#> 5   10.7     81   18.8
#> 6   10.8     83   19.7
#> 7   11.0     66   15.6
#> 8   11.0     75   18.2
#> 9   11.1     80   22.6
#> 10  11.2     75   19.9
#> 11  11.3     79   24.2
#> 12  11.4     76   21.0
#> 13  11.4     76   21.4
#> 14  11.7     69   21.3
#> 15  12.0     75   19.1
#> 16  12.9     74   22.2
#> 17  12.9     85   33.8
#> 18  13.3     86   27.4
#> 19  13.7     71   25.7
#> 20  13.8     64   24.9
#> 21  14.0     78   34.5
#> 22  14.2     80   31.7
#> 23  14.5     74   36.3
#> 24  16.0     72   38.3
#> 25  16.3     77   42.6
#> 26  17.3     81   55.4
#> 27  17.5     82   55.7
#> 28  17.9     80   58.3
#> 29  18.0     80   51.5
#> 30  18.0     80   51.0
#> 31  20.6     87   77.0
#>